Init visual Fire pass to reflect spread rate

+++ Ported/BucketBrigade/Assets/Scripts/Components/TuningData.cs
@@ -8,10 +8,10 @@ public struct TuningData : IComponentData
 {
 
     public Unity.Mathematics.int2 GridSize;
-    public int ValueIncreasePerTick;
-    public int MaxValue;
-    public int ValuePropagationPerTick;
-    public int ValueThreshold;
+    public float ValueIncreasePerTick;
+    public float MaxValue;
+    public float ValuePropagationPerTick;
+    public float ValueThreshold;
 
     public float ActorSpeed;
     public int BucketCapacity;
